initial commit
GPG with generated keys on the smartcard is basically working. Time for
the initial commit.

jcardsim moved to github package registry and doesn't offer the .jar
file download directly anymore. A package can be downloaded via the
following URL:
  https://maven.pkg.github.com/<org>/<repo>/com/<group>/<artifact>/<version>/<filename>
But one need an login token for it. Therefore, just use the latest
jcardsim.jar which is still available on an older revision of the repo.
If you want to use a newer jcardsim.jar you have to mantually download
the .jar on https://github.com/licel/jcardsim/packages/1650016 and
replace the file in lib/.

Signed-off-by: Michael Walle <michael@walle.cc>