NeoPIN: santizize length before checking
Prevent out of bounds access of the input buffer in case the stored PIN
is larger than the buffer. Limit the length to the input length. If this
happens the check function will fail. We must *not* return
SW_WRONG_LENGTH without checking the PIN first, because an attacker
could probe the PIN length this way.

Signed-off-by: Michael Walle <michael@walle.cc>