NeoKeyStore: fix borrowed Ciphers
Because Ciphers are associated with the private key, we need three
different Ciphers, one for each key type. Create three individual ones
and pass them to the key. This also means that one key can only do the
intended PSO, that is a signature key can only do signatures. Until we
implement the MSE operation, we are fine.

Signed-off-by: Michael Walle <michael@walle.cc>