NeoPGPApplet: support KDF
Add support for the key derivation function. The function is almost
transparent to the card. We need to serve an additonal KDF DO where the
KDF parameters are stored. All passwords will then transferred and saved
as hashes. That also means the card itself cannot verify the length of
the passwords anymore.

Because all passwords will have the same length this comes in handy to
workaround a bug on JCOP4 cards, where the OwnerPIN get broken if you
change the length to somehting larger than 15 bytes and then change it
back to a length smaller than 16 bytes. Due to that, it is even possible
to enable KDF by default.

Signed-off-by: Michael Walle <michael@walle.cc>